LPD-18491 Move some pages to use out of the journal
