+++ modules/test/playwright/pages/configuration-admin-web/InstanceSettingsPage.ts
@@ -5,7 +5,7 @@
 
 import {Page} from '@playwright/test';
 
-import {ApplicationsMenuPage} from '../../../pages/product-navigation-applications-menu/ApplicationsMenuPage';
+import {ApplicationsMenuPage} from '../product-navigation-applications-menu/ApplicationsMenuPage';
 
 export class InstanceSettingsPage {
 	readonly page: Page;
